requirements.txt
# MCP (Model Context Protocol) server
mcp>=1.0.0

# Web framework and ASGI server
fastapi>=0.104.0
uvicorn>=0.24.0

# HTTP client
httpx>=0.25.0

# CLI tools
click>=8.1.0
asyncclick>=8.1.0

# LangChain ecosystem
langchain-core>=0.1.0
langchain-ollama>=0.1.0
langgraph>=0.1.0
langchain-mcp-adapters>=0.1.0

# A2A framework (may need custom installation)
a2a-sdk<0.3

# Utilities
rich>=13.0.0
typing-extensions>=4.8.0

# Python version requirement
# python>=3.8